taxi_chap_backend/reservations/serializers.py
from rest_framework import serializers
from .models import Reservation, Vehicle

class ReservationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Reservation
        fields = '__all__'

class VehicleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vehicle
        fields = '__all__'
